requirements.txt
Babel==2.8.0
beautifulsoup4==4.9.1
colorama==0.4.3
defusedxml==0.6.0
diff-match-patch==20200713
Django==2.2.15
django-bootstrap-modal-forms==2.0.0
django-bootstrap4==2.2.0
django-countries==6.1.2
django-crum==0.7.7
django-import-export==2.3.0
django-phone-field==1.8.1
django-phonenumber-field==4.0.0
django-qrcode==0.3
django-widget-tweaks==1.4.8
et-xmlfile==1.0.1
importlib-metadata==1.7.0
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.4
phonenumbers==8.12.7
Pillow==7.2.0
pytz==2020.1
PyYAML==5.3.1
qrcode==6.1
simplejson==3.17.2
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
tablib==2.0.0
xlrd==1.2.0
xlwt==1.3.0
zipp==3.1.0
